requirements.txt
python-dotenv~=1.0.1
Flask~=3.0.3
redis~=5.0.8